Extract QEMU/virtiofsd logic into bcvk-qemu crate
Right now we have this unfortunate duality in that for `ephemeral`
we directly spawn qemu, but obviously the `libvirt` verb defers
all that to libvirt.

There are other use cases that effectively want to directly
control qemu, and while there are certainly *other* Rust
libraries wrapping qemu, ours is okay.

This makes our internals a crate that could be used as
a `git` dependency (no intention to publish to crates.io).

The kit crate now re-exports from bcvk-qemu to maintain backward
compatibility. A QemuConfigExt trait bridges kit's Format type with
bcvk-qemu's DiskFormat.

Tested by patching bootc PR #2018 to use bcvk-qemu via [patch.crates-io],
which builds successfully and uses spawn_virtiofsd_async() for the
anaconda test infrastructure.
